+++ src/idunno.AtProto/Repo/Models/ApplyWritesResponse.cs
@@ -1,8 +1,10 @@
 // Copyright (c) Barry Dorrans. All rights reserved.
 // Licensed under the MIT License.
 
+using System.Text.Json.Serialization;
+
 namespace idunno.AtProto.Repo.Models;
 
-internal sealed record ApplyWritesResponse(Commit Commit, IReadOnlyCollection<ApplyWritesResponseBase> Results)
+internal sealed record ApplyWritesResponse(Commit Commit, [property: JsonRequired] IReadOnlyCollection<ApplyWritesResponseBase> Results)
 {
 }