+++ src/idunno.Bluesky/Notifications/Model/ListNotificationsResponse.cs
@@ -1,8 +1,10 @@
 // Copyright (c) Barry Dorrans. All rights reserved.
 // Licensed under the MIT License.
 
+using System.Text.Json.Serialization;
+
 namespace idunno.Bluesky.Notifications.Model;
 
-internal sealed record ListNotificationsResponse(IList<NotificationResponse> Notifications, string? Cursor, bool? Priority, DateTimeOffset SeenAt)
+internal sealed record ListNotificationsResponse([property: JsonRequired] IList<NotificationResponse> Notifications, string? Cursor, bool? Priority, DateTimeOffset SeenAt)
 {
 }